Token naming refactor. (#5)
This commit performs a massive refactor of the system so that its naming and grammar
are more in line with the official grammar from OPA. That said, there are some areas
where the nature of Trieste as a multi-pass rewriter are not well-suited or
compatible with the grammar, and in these places I've made decisions to deviate in
order to ease implementation. However, this project now supports a proper subset
of the full language. Further, the rewrite will make future feature additions easier
to express, and also make tracking the difference between the two implementations
easier as we approach parity.

Signed-off-by: Matthew A Johnson <matjoh@microsoft.com>