Rules engine.
This commit replaces the rewrite logic used to compute rules with a C++ based rules engine.
It still operates on the AST, but instead of rewriting everything it constructs and resolves
the DAG which is headed by the query literals. In practical terms this greatly limits the
scope of the computation and allows for more targeted dependency traversal,  *e.g.*, just the
keys of an object that are needed, or only the specific rules in a module that are required by
to resolve the query. The result is a computational speedup and a firmer foundation for adding
further language features.

Signed-off-by: Matthew A Johnson <matjoh@microsoft.com>