Allow tests to see non-binding results
Sometimes, a test needs to assert results that are not bindings. Issue #109 shows a query of that form, where the intended output is a series of values with no bound names at all. Currently, the test suite will ignore those results and conclude there nothing is output by the query.

Conversely, sometimes rego-cpp outputs things it shouldn't, and we can't see it do that when those things aren't bindings. Issue #142 is an example of this, where the OPA test looks like it passes because all the bindings match, but if you look at the output of running rego-cpp manually, unwanted values are there.

This commit suggests a fix for the testing situation in both of these cases: assert both the binding and non-binding results. Existing tests become more powerful (and identify #142 correctly), and we can write tests for #109.